+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Ancestor items in context menus now stay tinted when their child menus are opened
 - A shortcut to open preferences (CMD/CTRL+,)
 - Support for lists in settings
+- Support for custom ids in select inputs in settings
 ### Changed
 - Some features have moved to the footer of the app window
 ### Fixed